chore: move sparse errors to `reth-execution-errors`
