+++ crates/trie/sparse/src/blinded.rs
@@ -1,7 +1,7 @@
 //! Traits and default implementations related to retrieval of blinded trie nodes.
 
-use crate::SparseTrieError;
 use alloy_primitives::{Bytes, B256};
+use reth_execution_errors::SparseTrieError;
 use reth_trie_common::Nibbles;
 
 /// Factory for instantiating blinded node providers.
