+++ crates/trie/sparse/src/lib.rs
@@ -6,7 +6,11 @@ pub use state::*;
 mod trie;
 pub use trie::*;
 
-mod errors;
-pub use errors::*;
-
 pub mod blinded;
+
+/// Re-export sparse trie error types.
+pub mod errors {
+    pub use reth_execution_errors::{
+        SparseStateTrieError, SparseStateTrieResult, SparseTrieError, SparseTrieResult,
+    };
+}
